+++ src/Mobs/Spider.java
@@ -4,7 +4,7 @@ package Mobs;
 import java.util.Random;
 
 public class Spider extends Monster {
-    public Spider(int hp, int dmg, int x, int y, String name, int giveXP, double level) {
+    public Spider(int hp, int dmg, double x, double y, String name, int giveXP, double level) {
         super(hp, dmg, x, y, name, giveXP, level);
 
     }
